requirements.txt
matplotlib~=3.2.1
Pillow~=7.1.2